Attempt to add pressure and temp stuff

+++ ROVController/src/Core/Engine.h
@@ -30,12 +30,12 @@ namespace Core
 
 		struct Temperature
 		{
-			float temp;
+			double temp;
 		};
 
 		struct Pressure
 		{
-			float pressure;
+			double pressure;
 		};
 
 		enum EventType
